refactor(billing): isolate Stripe naming behind BillingAccountRef and split use cases
Address review feedback:

- domain/entities/user_subscription.py no longer carries stripe_*
  field names. The provider-specific identifiers are wrapped in a new
  BillingAccountRef value object so the domain only sees opaque
  customer_id / subscription_id pairs.
- Postgres mapping translates the existing stripe_* columns to/from
  BillingAccountRef in PostgresUserSubscriptionRepository (DB schema
  unchanged).
- Drop unused IUserSubscriptionRepository.find_by_stripe_customer_id.
- Split manage_subscription.py into one-use-case-per-file modules:
  get_my_subscription, start_checkout, start_customer_portal,
  handle_stripe_webhook.
- Document FrontendUrls (what it's used for) and the Stripe-specific
  _with_session_id_token helper (why the placeholder is needed).
