+++ backend/domain/entities/user_subscription.py
@@ -3,26 +3,32 @@ from typing import Literal
 
 from pydantic import BaseModel, ConfigDict, Field
 
+from domain.value_objects.billing_account import BillingAccountRef
 from domain.value_objects.user_id import UserId
 
 SubscriptionPlan = Literal['free', 'paid']
 
 
 class UserSubscription(BaseModel):
+	"""A user's subscription record.
+
+	The optional ``billing`` field references an external billing provider
+	when the user has gone through Checkout at least once. Plain free users
+	have ``billing=None``.
+	"""
+
 	model_config = ConfigDict(frozen=False)
 
 	user_id: UserId
 	plan: SubscriptionPlan = 'free'
-	stripe_customer_id: str | None = None
-	stripe_subscription_id: str | None = None
-	current_period_end: datetime | None = None
-	cancel_at_period_end: bool = False
+	billing: BillingAccountRef | None = None
 	created_at: datetime = Field(default_factory=lambda: datetime.now(UTC))
 	updated_at: datetime = Field(default_factory=lambda: datetime.now(UTC))
 
 	def is_active_paid(self, *, now: datetime | None = None) -> bool:
 		if self.plan != 'paid':
 			return False
-		if self.current_period_end is None:
+		if self.billing is None:
+			# Plan is paid but no billing reference — treat as active.
 			return True
-		return (now or datetime.now(UTC)) < self.current_period_end
+		return self.billing.is_within_period(now=now)
