+++ backend/domain/value_objects/__init__.py
@@ -1,4 +1,5 @@
 from .arxiv_paper_id import ArxivPaperId
+from .billing_account import BillingAccountRef
 from .blog_paper_id import BlogPaperId, InvalidBlogPaperIdError
 from .blog_source_type import BlogSourceType, InvalidBlogSourceTypeError
 from .compile_setting import CompileSetting
@@ -11,6 +12,7 @@ from .user_role import UserRole
 
 __all__ = [
 	'ArxivPaperId',
+	'BillingAccountRef',
 	'BlogPaperId',
 	'BlogSourceType',
 	'CompileSetting',
