+++ backend/infrastructure/postgres/repositories/postgres_user_subscription_repository.py
@@ -8,25 +8,39 @@ from sqlmodel import col
 
 from domain.entities.user_subscription import UserSubscription
 from domain.repositories import IUserSubscriptionRepository
+from domain.value_objects.billing_account import BillingAccountRef
 from domain.value_objects.user_id import UserId
 
 from ..models import UserSubscriptionModel
 
 
 class PostgresUserSubscriptionRepository(IUserSubscriptionRepository):
+	"""Maps ``UserSubscription`` to the ``user_subscription`` table.
+
+	The table still stores Stripe-specific column names (``stripe_*``) since
+	Stripe is the only provider we currently use. The mapping here translates
+	those columns into the provider-agnostic ``BillingAccountRef`` value
+	object on the domain side.
+	"""
+
 	def __init__(self, session: AsyncSession):
 		self._session = session
 		self._logger = getLogger(__name__)
 
 	def _to_entity(self, row: UserSubscriptionModel) -> UserSubscription:
 		plan = row.plan if row.plan in ('free', 'paid') else 'free'
+		billing: BillingAccountRef | None = None
+		if row.stripe_customer_id is not None:
+			billing = BillingAccountRef(
+				customer_id=row.stripe_customer_id,
+				subscription_id=row.stripe_subscription_id,
+				current_period_end=row.current_period_end,
+				cancel_at_period_end=row.cancel_at_period_end,
+			)
 		return UserSubscription(
 			user_id=UserId(row.user_id),
 			plan=plan,  # type: ignore[arg-type]
-			stripe_customer_id=row.stripe_customer_id,
-			stripe_subscription_id=row.stripe_subscription_id,
-			current_period_end=row.current_period_end,
-			cancel_at_period_end=row.cancel_at_period_end,
+			billing=billing,
 			created_at=row.created_at,
 			updated_at=row.updated_at,
 		)
@@ -39,25 +53,16 @@ class PostgresUserSubscriptionRepository(IUserSubscriptionRepository):
 		row = result.scalars().first()
 		return self._to_entity(row) if row is not None else None
 
-	async def find_by_stripe_customer_id(
-		self, stripe_customer_id: str
-	) -> UserSubscription | None:
-		stmt = select(UserSubscriptionModel).where(
-			col(UserSubscriptionModel.stripe_customer_id) == stripe_customer_id
-		)
-		result = await self._session.execute(stmt)
-		row = result.scalars().first()
-		return self._to_entity(row) if row is not None else None
-
 	async def upsert(self, subscription: UserSubscription) -> UserSubscription:
 		now = datetime.now(UTC)
+		billing = subscription.billing
 		values = {
 			'user_id': subscription.user_id.root,
 			'plan': subscription.plan,
-			'stripe_customer_id': subscription.stripe_customer_id,
-			'stripe_subscription_id': subscription.stripe_subscription_id,
-			'current_period_end': subscription.current_period_end,
-			'cancel_at_period_end': subscription.cancel_at_period_end,
+			'stripe_customer_id': billing.customer_id if billing else None,
+			'stripe_subscription_id': billing.subscription_id if billing else None,
+			'current_period_end': billing.current_period_end if billing else None,
+			'cancel_at_period_end': billing.cancel_at_period_end if billing else False,
 			'created_at': subscription.created_at or now,
 			'updated_at': now,
 		}
